fix/fixUpButtom
