+++ src/pages/edit/style1.css
@@ -1,0 +1,106 @@
+
+@import url('https://fonts.googleapis.com/css2?family=Outfit:wght@100..900&display=swap');
+*{
+    margin: 0;
+    padding: 0;
+}
+
+.student{
+    width: 979px;
+    height: 417px;
+    flex-shrink: 0;
+    border: 1px solid #E30613;
+    display: flex;
+    align-items: center;
+    gap: 55px;
+    padding-left: 80px;
+}
+
+.student_description{
+    display: flex;
+    flex-direction: column;
+
+}
+
+.student_description #snome{
+    color: #000;
+    font-size: 15px;
+    padding: 10px;
+    border-radius: 5px;
+    text-align: center;
+    background: rgb(217,217,217,0.25); 
+    border: none;  
+}
+
+.student_description #sdescricao{
+    color: #000;
+    border-radius: 5px;
+    border: none; 
+    width: 450px;
+    height: 120px;
+    background: rgb(217,217,217,0.25);  
+}
+
+#des{
+margin-top: 10px;
+}
+
+.student_description button{
+    background-color: red;
+    border: none;
+    width: 150px;
+    height: 30px;
+    color: white;
+    font-size: 20px;
+    align-items: center;
+    font-family: "Outfit", sans-serif;
+    justify-content: center;
+    border-radius: 20px;
+    font-weight: 500;
+    margin-top: 55px;
+}
+
+::placeholder,label{
+    color: rgb(0,0,0,0.3);
+    font-size: 15px;
+    font-family: "Inter", sans-serif; 
+}
+
+.student_info{
+    display: flex;
+    flex-direction: column;
+ 
+}
+
+.student_info img{
+    width: 225px;
+    height: 230px;
+    flex-shrink: 0;
+}
+
+.student_info #simg{
+    color: #000;
+    font-size: 15px;
+    margin: 5px 10px;
+    padding: 90px 10px;
+    border-radius: 5px;
+    border: none;
+    text-align: center;
+    background: rgb(217,217,217,0.25); 
+}
+
+.student_info button{
+    background-color: red;
+    border: none;
+    width: 150px;
+    height: 30px;
+    color: white;
+    font-size: 20px;
+    align-items: center;
+    font-family: "Outfit", sans-serif;
+    justify-content: center;
+    border-radius: 20px;
+    font-weight: 500;
+    width: 225px;
+    margin-top: 10px;
+}
